refactor: extract functions & frameworks to utils
