+++ ramon/.eslintrc
@@ -1,0 +1,30 @@
+{
+  "env": {
+    "browser": true,
+    "es2021": true
+  },
+  "extends": [
+    "plugin:react/recommended",
+    "plugin:react/jsx-runtime",
+    "plugin:prettier/recommended",
+    "plugin:jsx-a11y/recommended",
+    "standard"
+  ],
+  "overrides": [],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "plugins": ["react", "prettier", "jsx-a11y"],
+  "rules": {
+    "react/prop-types": "off",
+    "space-before-function-paren": "off",
+    "prettier/prettier": 2
+  },
+  "globals": {
+    "expect": "writable",
+    "describe": "writable",
+    "it": "writable",
+    "test": "writable"
+  }
+}
