+++ ramon/.prettierrc
@@ -1,0 +1,7 @@
+{
+  "semi": false,
+  "trailingComma": "none",
+  "singleQuote": true,
+  "printWidth": 80,
+  "endOfLine": "auto"
+}
