+++ ramon/src/main.jsx
@@ -1,0 +1,12 @@
+import { ChakraProvider } from '@chakra-ui/react'
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import App from './App'
+
+ReactDOM.createRoot(document.getElementById('root')).render(
+  <React.StrictMode>
+    <ChakraProvider>
+      <App />
+    </ChakraProvider>
+  </React.StrictMode>
+)
